calendar-service/requirements.txt
fastapi==0.115.0 
uvicorn==0.30.6 
sqlalchemy==2.0.35 
psycopg2-binary==2.9.9 
pydantic==2.9.2 
apscheduler==3.10.4 
python-dotenv==1.0.1
websockets==15.0.1
python-jose[cryptography]==3.3.0